+++ src/GuppyDataExplorer/ExplorerTableOne/RiskTable.jsx
